UI tweaks to fuel/silo display
It was confusing that sometimes the volumes displayed were the volume of
the contents, and sometimes the volume of the free space, depending on
whether the silo was an input or output. This makes the meaning
consistent by just showing both. It also includes quantities in addition
to volumes, which is helpful when e.g. shift-dragging from a stack into
your hauler.

The meter bridges the conceptual gap between the free/filled numbers and
the input/output arrow indicator. Without this, it would be easy to
think that a down (input) arrow indicates quantity being transfered from
the upper (free) line to the lower (filled) line, but this is the
opposite of reality. The meter matches nicely both with the numbers to
the left (the upper line describes the upper/empty part of the meter)
and the right (a down arrow indicates that the meter is falling).

I considered trying to fit a percentage indicator in somewhere, but I
don't think it would provide much value. The meter gives a rough
indication, and who cares about the exact percentage?

The meter css feels a bit sketch (rotation + negative margin to offset
the still-applied width attribute). Maybe it would be better to roll our
own meter UI with colored divs. But this works well enough in both
Chrome and Firefox that it's probably fine (though they seem to treat
the combination of rotation & margin differently).